# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Golang Quickstart using Gin Gonic",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/api/v1/airline/list": {
            "get": {
                "description": "Get list of Airlines. Optionally, you can filter the list by Country",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline collection"
                ],
                "summary": "Get Airlines by Country",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by country\u003cbr\u003eExample: France, United Kingdom, United States",
                        "name": "country",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of airlines to return (page size).\u003cbr\u003eExample: 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of airlines to skip (for pagination).\u003cbr\u003eExample: 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Airline"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/airline/to-airport": {
            "get": {
                "description": "Get Airlines flying to specified destination Airport",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline collection"
                ],
                "summary": "Get Airlines Flying to Airport",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination airport\u003cbr\u003eExample : SFO, JFK, LAX",
                        "name": "airport",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Number of airlines to return (page size)\u003cbr\u003eDefault value : 10",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of airlines to skip (for pagination)\u003cbr\u003eDefault value : 0",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Airline"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/airline/{id}": {
            "get": {
                "description": "Get Airline with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline collection"
                ],
                "summary": "Get Airline Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airline ID like airline_10",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Airline"
                        }
                    },
                    "404": {
                        "description": "Airline Document ID not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update Airline with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline collection"
                ],
                "summary": "Update Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airline ID like airline_10",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updates document",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Airline"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Airline"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create Airline with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline collection"
                ],
                "summary": "Insert Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airline ID like airline_10",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data to create a new document",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Airline"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Airline"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Airline document already exists"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete Airline with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Airline collection"
                ],
                "summary": "Delete Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Airline ID like airline_10",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Airline deleted"
                    },
                    "404": {
                        "description": "Airline Document ID Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.Airline": {
            "type": "object",
            "required": [
                "country",
                "icao",
                "name"
            ],
            "properties": {
                "callsign": {
                    "type": "string",
                    "example": "SAF"
                },
                "country": {
                    "type": "string",
                    "example": "United States"
                },
                "iata": {
                    "type": "string",
                    "example": "SA"
                },
                "icao": {
                    "type": "string",
                    "example": "SAF"
                },
                "name": {
                    "type": "string",
                    "example": "SampleName"
                }
            }
        },
...
    }
}